move more stuff to lib.rs
This fixes rustdoc.

In addition, this enables doc-tests.
